Refactored project, added template base
