+++ src/main/java/fi/hh/swd4t/zone/domain/QuestionRepository.java
@@ -1,10 +1,7 @@
-package fi.hh.swd4t.zone.fi.hh.swd4t.zone.dao;
+package fi.hh.swd4t.zone.domain;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.transaction.annotation.Transactional;
-
-
-import fi.hh.swd4t.zone.fi.hh.swd4t.zone.web.Question;
 
 public interface QuestionRepository extends CrudRepository<Question, Long> {
 	@Transactional
