+++ src/main/resources/templates/addbook.html
@@ -1,0 +1,51 @@
+<!DOCTYPEhtml>
+<html xmlns="http://www.w3.org/1999/xhtml"
+		xmlns:th="http://www.thymeleaf.org">
+<head>
+	<title>Form page for adding books</title>
+	<link type="text/css" rel="stylesheet" href="css/bootstrap.min.css" th:href="@{css/bootstrap.min.css}" />
+</head>
+	<body>
+	<div class="container">
+		<h1>Add book</h1>
+		<p th:text="${toiminto}"></p>
+	<div>
+			<form th:object="${book}" th:action="@{/addbook}" action="#" method="post">
+			<div class="form-group">
+			<label for="title">Title</label>
+			<input type="text" id="title" th:field="*{title}" class="form-control"/>
+			</div>
+			
+			<div class="form-group">
+			<label for="author">Author</label>
+			<input type="text" id="author" th:field="*{author}" class="form-control"/>
+			</div>
+			
+			<div class="form-group">
+			<label for="year">Year</label>
+			<input type="text" th:field="*{year}" class="form-control"/>
+			</div>
+			
+			<div class="form-group">
+			<label for="price">Price</label>
+			<input type="text" th:field="*{price}" class="form-control"/>
+			</div>
+			
+			<div class="form-group">
+			<label for="isbn">ISBN</label>
+			<input type="text" th:field="*{isbn}" class="form-control"/>
+			</div>
+			
+			<label for="catlist">Category</label>
+			<select id="catlist" th:field="*{category}" class="form-control">
+				<option th:each="category : ${categories}" 
+				th:value="${category.id}" th:text="${category.name}"></option>
+			</select>
+			
+			<input type="submit" value="Save" class="btn btn-primary"></input>
+		</form>
+	</div>
+	<a href="/booklist">Back to list</a>
+	</div>
+	</body>
+</html>
